lct_ai_service_package/requirements.txt
# LCT AI Service - Обновленные зависимости
# Поддержка новых требований: Hadoop + Spark + AI + Kafka

# Основные зависимости
pandas==2.1.4
requests==2.31.0
python-dotenv==1.0.0
numpy==1.24.3

# AI и LLM интеграция
openai==1.3.7

# База данных
psycopg2-binary==2.9.9
sqlalchemy==2.0.23

# Kafka интеграция
kafka-python==2.0.2

# Hadoop и Big Data
pyspark==3.4.0
py4j==0.10.9.7
hdfs3==0.3.1

# Веб и API
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0

# Утилиты
pyyaml==6.0.1
click==8.1.7
colorama==0.4.6
aiofiles==23.2.1
httpx==0.25.2

# Тестирование
pytest==7.4.3

# Логирование и мониторинг
structlog==23.2.0